Add response and status on DELETE request

+++ ch3/src/main/kotlin/com/microservices/ch3/CustomerController.kt
@@ -25,8 +25,13 @@ class CustomerController {
     }
 
     @DeleteMapping(value = ["/customer/{id}"])
-    fun deleteCustomer(@PathVariable id : Int) {
-        customerService.deleteCustomer(id)
+    fun deleteCustomer(@PathVariable id : Int): ResponseEntity<Unit> {
+        var status = HttpStatus.NOT_FOUND
+        if (customerService.getCustomer(id) != null) {
+            customerService.deleteCustomer(id)
+            status = HttpStatus.OK
+        }
+        return ResponseEntity(Unit, status)
     }
 
     @PutMapping(value = ["/customer/{id}"])
